+++ udata_front/frontend/helpers.py
@@ -17,6 +17,7 @@ from . import front
 
 from udata.core.dataset.apiv2 import dataset_fields
 from udata.core.dataset.models import Dataset
+from udata.core.dataservices.models import Dataservice
 from udata.core.organization.models import Organization
 from udata.core.organization.constants import (
     ASSOCIATION,
@@ -419,6 +420,11 @@ def to_api_format(data):
 
 def to_dataset_api_format(dataset):
     return marshal(dataset, dataset_fields)
+
+
+@front.app_template_filter()
+def to_dataservice_api_format(dataservice):
+    return marshal(dataservice, Dataservice.__read_fields__)
 
 
 @front.app_template_filter()