+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Current (in progress)
 
+- add dataservices to dataset page [#476](https://github.com/datagouv/udata-front/pull/476)
+- show contact point in dataset and dataservice page [#479](https://github.com/datagouv/udata-front/pull/479)
 - Show organization type [#472](https://github.com/datagouv/udata-front/pull/472)
 
 ## 5.1.2 (2024-08-01)